// src/main/resources/angular/src/app/api/model/vacationDay.ts
/**
 * Smou API
 * This is the Mobility as a Service API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: avalls@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface VacationDay {
    vacationId: number;
    vacationDay: string;
    description?: string;
}